Basics completed
The bare bones of DCnet actually work now. Here is an example:

[DCnet] station -m s1 s2 s3
[DCnet] station -c s1 s2 s1 s3 s2 s3
[DCnet] station s1 send somethingsomethi
[DCnet] station s2 send somethingelsebla
[DCnet] station s3 send somethingevendif
[DCnet] station s1 read
1: somethingsu{nrmn
[DCnet] station s2 read
1: somethingsu{nrmn
[DCnet] station s3 read
1: somethingsu{nrmn

As you can see, those parts that have been broadcasted an odd number of times have remained unchanged, but especially the end of the message is corrupted. Also, the 'e' at position 10 is transmitted 2 times, and thus cancelles out. Instead we see the the 's' in the output that has only been submitted once.

Command reference update will follow shortly.
For now: if you want to be silent for this round, use
	station sX send -s