Deadlock prevention and message buffering
Changed message sending behaviour so that output is broadcasted even though the number of required connections is not reached. However, this output will always be empty as long as there are too little connections.
Once the minimum number of connections is reached, the station will broadcast output that contains all messages that have built up in the past.